feat: add InvokeFn cmd
Capture only works for one arg at a time.
Sometimes we need to capture multiple args at a time (to assert an invariant) -- e.g. buffer ptr and length.
Rather than adding to the Capture action, use invoke instead to invoke a method with *all* args.
This is inspired by a similar function in googlemock.

Note that we needed to change skeletal to non constexpr but this does not seem to have a perf penalty as of now